+++ config/src/Config.ts
@@ -1,0 +1,7 @@
+import type { ActivityDefinition } from "./ActivityDefinition";
+
+export type Config<T extends ActivityDefinition<string>> = {
+  activities: T[];
+  transitionDuration: number;
+  initialActivity?: () => T["name"];
+};