+++ config/src/defineConfig.ts
@@ -1,0 +1,9 @@
+import type { ActivityDefinition } from "./ActivityDefinition";
+import type { Config } from "./Config";
+
+export function defineConfig<
+  ActivityName extends string,
+  Activity extends ActivityDefinition<ActivityName>,
+>(config: Config<Activity>) {
+  return config;
+}