+++ demo/src/activities/Main.loader.ts
@@ -1,0 +1,76 @@
+export function mainLoader() {
+  return {
+    cards: [
+      {
+        articleId: "02542470",
+        price: 41,
+        title: "Master",
+        region: "Nagevan",
+        daysAgo: 4,
+      },
+      {
+        articleId: "11257089",
+        price: 24,
+        title: "Wild",
+        region: "Inguima",
+        daysAgo: 4,
+      },
+      {
+        articleId: "08407137",
+        price: 42,
+        title: "Universe",
+        region: "Litenego",
+        daysAgo: 4,
+      },
+      {
+        articleId: "32979422",
+        price: 12,
+        title: "Private",
+        region: "Umumtaw",
+        daysAgo: 6,
+      },
+      {
+        articleId: "37998208",
+        price: 3,
+        title: "Harbor",
+        region: "Gubdidgi",
+        daysAgo: 3,
+      },
+      {
+        articleId: "01695878",
+        price: 1,
+        title: "Valuable",
+        region: "Jumjelewu",
+        daysAgo: 1,
+      },
+      {
+        articleId: "09792471",
+        price: 31,
+        title: "Also",
+        region: "Salhega",
+        daysAgo: 1,
+      },
+      {
+        articleId: "23939055",
+        price: 49,
+        title: "Ever",
+        region: "Jaifuup",
+        daysAgo: 9,
+      },
+      {
+        articleId: "94689745",
+        price: 26,
+        title: "Production",
+        region: "Idcipwel",
+        daysAgo: 3,
+      },
+      {
+        articleId: "49322156",
+        price: 35,
+        title: "Chest",
+        region: "Ajapaktar",
+        daysAgo: 7,
+      },
+    ],
+  };
+}