+++ integrations/react/src/activity/ActivityComponentType.tsx
@@ -1,2 +1,0 @@
-export type ActivityComponentType<T extends { [K in keyof T]: any } = {}> =
-  React.ComponentType<{ params: T }>;